+++ packages/amplify-category-geo/src/__tests__/commands/geo/add.test.ts
@@ -1,4 +1,4 @@
-import { $TSContext } from 'amplify-cli-core';
+import { $TSContext, stateManager, $TSObject } from 'amplify-cli-core';
 import { addResource } from '../../../provider-controllers/index';
 import { ServiceName } from '../../../service-utils/constants';
 import { run } from '../../../commands/geo/add';
@@ -13,6 +13,10 @@ jest.mock('../../../provider-controllers/index');
 describe('add command tests', () => {
     const provider = 'awscloudformation';
     let mockContext: $TSContext;
+    // construct mock amplify meta
+    const mockAmplifyMeta: $TSObject = {
+        providers: {}
+    };
         
     beforeEach(() => {
         jest.clearAllMocks();
@@ -23,6 +27,10 @@ describe('add command tests', () => {
             },
             amplify: {}
         } as unknown) as $TSContext;
+        mockAmplifyMeta.providers[provider] = {
+            Region: 'us-west-2'
+        };
+        stateManager.getMeta = jest.fn().mockReturnValue(mockAmplifyMeta);
     });
 
     it('add resource workflow is invoked for map service', async() => {
@@ -45,5 +53,16 @@ describe('add command tests', () => {
         await run(mockContext);
 
         expect(mockAddResource).toHaveBeenCalledWith(mockContext, service);
+    });
+
+    it('add resource workflow is not invoked for unsupported region', async() => {
+        mockAmplifyMeta.providers[provider] = {
+            Region: 'eu-west-2'
+        };
+        stateManager.getMeta = jest.fn().mockReturnValue(mockAmplifyMeta);
+
+        await run(mockContext);
+
+        expect(mockAddResource).toBeCalledTimes(0);
     });
 });