+++ packages/amplify-category-geo/src/__tests__/commands/geo/remove.test.ts
@@ -1,4 +1,4 @@
-import { $TSContext } from 'amplify-cli-core';
+import { $TSContext, stateManager, $TSObject } from 'amplify-cli-core';
 import { removeResource } from '../../../provider-controllers';
 import { ServiceName } from '../../../service-utils/constants';
 import { run } from '../../../commands/geo/remove';
@@ -18,6 +18,10 @@ jest.mock('../../../provider-controllers');
 describe('remove command tests', () => {
     const provider = 'awscloudformation';
     let mockContext: $TSContext;
+    // construct mock amplify meta
+    const mockAmplifyMeta: $TSObject = {
+        providers: {}
+    };
     
     beforeEach(() => {
         jest.clearAllMocks();
@@ -28,6 +32,10 @@ describe('remove command tests', () => {
             },
             amplify: {}
         } as unknown) as $TSContext;
+        mockAmplifyMeta.providers[provider] = {
+            Region: 'us-west-2'
+        };
+        stateManager.getMeta = jest.fn().mockReturnValue(mockAmplifyMeta);
     });
 
     it('remove resource workflow is invoked for map service', async() => {