+++ packages/amplify-cli-core/src/index.ts
@@ -24,6 +24,8 @@ export * from './cliGetCategories';
 export * from './cliRemoveResourcePrompt';
 export * from './cliViewAPI';
 export * from './hooks';
+export * from './cliViewAPI';
+export * from './customPoliciesUtils'
 
 // Temporary types until we can finish full type definition across the whole CLI
 
@@ -56,7 +58,7 @@ export type ResourceName = string;
 
 export type IContextPrint = {
   /**
-   * @deprecated Use printer.info from amplify-prommpts instead
+   * @deprecated Use printer.info from amplify-prompts instead
    */
   info: (message: string) => void;
   /**
@@ -72,7 +74,7 @@ export type IContextPrint = {
    */
   error: (message: string) => void;
   /**
-   * @deprecated Use printer.success from amplify-prommpts instead
+   * @deprecated Use printer.success from amplify-prompts instead
    */
   success: (message: string) => void;
   /**
@@ -253,7 +255,7 @@ interface AmplifyToolkit {
     filteredResources?: { category: string; resourceName: string }[],
   ) => $TSAny;
   storeCurrentCloudBackend: () => $TSAny;
-  readJsonFile: () => $TSAny;
+  readJsonFile: (fileName: string) => $TSAny;
   removeDeploymentSecrets: (context: $TSContext, category: string, resource: string) => void;
   removeResource: (
     context: $TSContext,
@@ -319,9 +321,7 @@ interface AmplifyToolkit {
   leaveBreadcrumbs: (category: string, resourceName: string, breadcrumbs: unknown) => void;
   readBreadcrumbs: (category: string, resourceName: string) => $TSAny;
   loadRuntimePlugin: (context: $TSContext, pluginId: string) => Promise<$TSAny>;
-  getImportedAuthProperties: (
-    context: $TSContext,
-  ) => {
+  getImportedAuthProperties: (context: $TSContext) => {
     imported: boolean;
     userPoolId?: string;
     authRoleArn?: string;