+++ packages/amplify-graphql-transformer-core/src/transformer-context/resource-helper.ts
@@ -1,6 +1,7 @@
 import { GraphQLAPIProvider, TransformerResourceHelperProvider } from '@aws-amplify/graphql-transformer-interfaces';
 import { CfnParameter } from '@aws-cdk/core';
 import { StackManager } from './stack-manager';
+import md5 from 'md5';
 
 export class TransformerResourceHelper implements TransformerResourceHelperProvider {
   private api?: GraphQLAPIProvider;
@@ -9,7 +10,7 @@ export class TransformerResourceHelper implements TransformerResourceHelperProvi
   constructor(private stackManager: StackManager) {}
   generateTableName = (modelName: string): string => {
     if (!this.api) {
-      throw new Error('API not initalized');
+      throw new Error('API not initialized');
     }
     this.ensureEnv();
     const env = (this.stackManager.getParameter('env') as CfnParameter).valueAsString;
@@ -26,6 +27,19 @@ export class TransformerResourceHelper implements TransformerResourceHelperProvi
     const env = (this.stackManager.getParameter('env') as CfnParameter).valueAsString;
     const apiId = this.api!.apiId;
     return `${baseName}-${apiId}-${env}`;
+  };
+
+  public generateIAMRoleName = (name: string): string => {
+    if (!this.api) {
+      throw new Error('API not initialized');
+    }
+    this.ensureEnv();
+    const env = (this.stackManager.getParameter('env') as CfnParameter).valueAsString;
+    const apiId = this.api!.apiId;
+
+    // 38 = 26(apiId) + 10(env) + 2(-)
+    const shortName = name.slice(0, 64 - 38 - 6) + md5(name).slice(0, 6);
+    return `${shortName}-${apiId}-${env}`; // max of 64.
   };
 
   bind(api: GraphQLAPIProvider) {