+++ packages/amplify-graphql-transformer-interfaces/src/transformer-context/resource-resource-provider.ts
@@ -1,6 +1,6 @@
 export interface TransformerResourceHelperProvider {
   generateTableName(modelName: string): string;
-  generateRoleName(baseName: string): string;
+  generateIAMRoleName(name: string): string;
   registerModelToTableNameMaping(modelName: string, tableName: string): void;
   getTableBaseName(modelName: string): string;
 }