+++ packages/amplify-provider-awscloudformation/src/graphql-transformer/transform-graphql-schema.ts
@@ -32,6 +32,8 @@ import { Template } from '@aws-amplify/graphql-transformer-core/lib/config/proje
 import { AmplifyCLIFeatureFlagAdapter } from '../utils/amplify-cli-feature-flag-adapter';
 import { JSONUtilities } from 'amplify-cli-core';
 import { V2WasTransformer } from './was-transformer';
+import { searchablePushChecks } from '../transform-graphql-schema';
+import { ResourceConstants } from 'graphql-transformer-common';
 
 const API_CATEGORY = 'api';
 const STORAGE_CATEGORY = 'storage';
@@ -277,6 +279,7 @@ export async function transformGraphQLSchema(context, options) {
   // Check for common errors
   const directiveMap = collectDirectivesByTypeNames(project.schema);
   warnOnAuth(context, directiveMap.types);
+  searchablePushChecks(context, directiveMap.types, parameters[ResourceConstants.PARAMETERS.AppSyncApiName]);
 
   const transformerListFactory = getTransformerFactory(context, resourceDir);