+++ packages/amplify-app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.14](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@3.0.13...amplify-app@3.0.14) (2021-10-06)
+
+**Note:** Version bump only for package amplify-app
+
+
+
+
+
+## [3.0.13](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@3.0.12...amplify-app@3.0.13) (2021-09-27)
+
+**Note:** Version bump only for package amplify-app
+
+
+
+
+
 ## [3.0.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@3.0.11...amplify-app@3.0.12) (2021-09-18)
 
 **Note:** Version bump only for package amplify-app