+++ packages/amplify-category-predictions/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.14](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.13...amplify-category-predictions@2.9.14) (2021-10-06)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
+## [2.9.13](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.12...amplify-category-predictions@2.9.13) (2021-09-27)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
 ## [2.9.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.9.11...amplify-category-predictions@2.9.12) (2021-09-18)
 
 **Note:** Version bump only for package amplify-category-predictions