+++ packages/amplify-category-xr/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.23](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-xr@2.8.22...amplify-category-xr@2.8.23) (2021-10-06)
+
+**Note:** Version bump only for package amplify-category-xr
+
+
+
+
+
+## [2.8.22](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-xr@2.8.21...amplify-category-xr@2.8.22) (2021-09-27)
+
+**Note:** Version bump only for package amplify-category-xr
+
+
+
+
+
 ## [2.8.21](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-xr@2.8.20...amplify-category-xr@2.8.21) (2021-09-18)
 
 **Note:** Version bump only for package amplify-category-xr