+++ packages/amplify-python-function-runtime-provider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.13](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-runtime-provider@1.9.12...amplify-python-function-runtime-provider@1.9.13) (2021-10-06)
+
+**Note:** Version bump only for package amplify-python-function-runtime-provider
+
+
+
+
+
+## [1.9.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-runtime-provider@1.9.11...amplify-python-function-runtime-provider@1.9.12) (2021-09-27)
+
+**Note:** Version bump only for package amplify-python-function-runtime-provider
+
+
+
+
+
 ## [1.9.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-runtime-provider@1.9.10...amplify-python-function-runtime-provider@1.9.11) (2021-09-18)
 
 **Note:** Version bump only for package amplify-python-function-runtime-provider