+++ packages/amplify-util-mock/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.34.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.7...amplify-util-mock@3.34.8) (2021-10-06)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [3.34.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.6...amplify-util-mock@3.34.7) (2021-10-01)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [3.34.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.5...amplify-util-mock@3.34.6) (2021-09-27)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
 ## [3.34.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.34.4...amplify-util-mock@3.34.5) (2021-09-18)
 
 **Note:** Version bump only for package amplify-util-mock